Adds initial PEG, basic recognizer API, and Error/Result APIs.
Adds `pest`/`pest_derive` dependencies and a very basic PEG definition
that is just a stream of PartiQL keywords.  This implementation is just
to flesh out some of the usage of `pest`, namely the derivation macro
and its error APIs.

This commit exposes a function `recognize_partiql` that runs the parser
over an input slice and returns `()` on success or fails on invalid
syntax.  Given that the syntax is just a stream of keywords, this is not
particularly interesting right now.

Adds the following modules:
* `result` contains the main `ParserError` enum, `Position`, and
  `ParserResult` alias.
* `prelude` exports most of `result` and will be used in the future for
  things like traits as well.
* `peg` is the module that hosts the low-level Pest PEG parser and the
  grammar definition file.

Other notes:
* Adds `From` implementation for `pest::error::Error` for `ParserError`
  to do basic translation of the relatively obtuse Pest error to
  something a bit more friendly.
* Updates README around versioning.
* Changes stub tests to just be no-op for the non-parser packages.
* Adds dependency for `thiserror` for less boilerplate error
  definitions.
* Adds dev dependency for `rstest` for table driven testing.