Changes construction of `LineAndColumn` to return `ParserResult`.
The rationale is that even though there are some aspects of this that
we could assert as internal logic problems, we rely on external APIs
to construct this location and we should not panic as a result of those.

Refactors relevant APIs/tests to use this new return result.

Adds `TryFrom` and `TryInto` into the prelude.

This also cleans up field access for the `LineAndColumn` struct.