Adds support for string literals and identifiers in `Scanner`. (#15)
Refactors the PEG to make each keyword and explicit rule that has the
form `KeywordName_`.  This allows us to easily refer to them in rules
without having to have magic strings.  The rationale for the `_` suffix
is to make sure it doesn't collide with the other grammar rules and not
have a wordy suffix like `Keyword` on every rule.

Also:
* Adds case folding to keywords.
* Makes error cases in the scanner tests less brittle.
* Adds `Token` to the `Query` rule and updates tests for
  `recognize_partiql`.