Streamlines the `Scanner` Tests. (#25)
Previously, the `Scanner` tests required a lot of manual position
expectations which was tedious to write.  This replaces that with a test
harness that can do the math from chunks of input that the test writer
defines as whitespace or token boundaries.  This also adds the macro
`scanner_test_case!` which makes it so that one can use the harness with
a friendly syntax. The old tests are ported to the new macro.

Also:
* Adds assertions to the internal `LineAndColumn::at`.
* Adds convenience constructor around `Content`.

Resolves #21.