Fixes an ambiguity with identifiers and keywords. (#32)
Previously, `SELECTY` would be parsed as keyword `SELECT` identifier
`Y`.  This fixes the negative look-ahead assertion for keywords to
make sure a non-quoted identifier character follows and similarly
adds a look-ahead assertion for the prefix of identifiers to not be
keywords (which includes the previous assertion).

Also fixes a unit test name for identifiers.