Adds Pest Grammar to Ion Conversion
The purpose of this commit is the start of a tool to generate Ion
formatted ASTs from Pest grammars, particularly that of PartiQL.
We could then easily process the grammar outside of Rust in scripts that
could be used to generate things like TextMate grammars for IDE support
or LaTeX fragments for the specification document.

Adds the `pestion` crate to provide a simple trait `PestToElement` and
implementations over the Pest AST and `&str` syntax definitions to
create Ion `Element` serialized forms.

Resolves #35.