Adds fixes to avoid heap allocations
As per feedback from @zslayton.

Renames `PestToElement` to `TryPestToElement` which captures the
fallible conversion (e.g. from `str` slices) and makes `PestToElement`
be an infallible version of the API.  This is a more correct factoring,
but also needed to avoid the intermediate copy for `Vec<AstNode>`
conversion.

Changes `AstRule` structure generation to use an array as it has a fixed
number of fields.

Adds `smallvec` to make it easier to model dynamically sized yet stack
allocated fields for sub-structures for `Expr` conversion.

Adjusts the doc test to be more illustrative.