Adds `pest2ion` command (#44)
Adds a basic CLI driver over the `pest-ion` crate.  This driver supports
setting the input file (defaulting to STDIN), an output file (defaulting
to STDOUT), and flags around the Ion data format.

An explicit TODO is to figure out a good way to test this probably using
integration tests and spawning the command. (See #42)

Resolves #36.