Adds CLI tests for `pest2ion` (#45)
Uses `assert_cmd` and `rstest` to test the different permutations of the
`pest2ion` command with real arguments.  This includes testing
STDIN/STDOUT defaults as well as file arguments using a temporary
directory per test case.  Each format mode is also permuted.

Note that this does not test the "long" argument names, if we care to,
we can always add that for completeness.

Resolves #42.