Adds `pest2ion` CLI test for PartiQL Grammar
Adds a `TestCase` struct to encapsulate the test case parameters for
testing the CLI.  Makes a simple conversion from `&'static str` pair
which is for when you have simple in/out type of tests.  Adds function
to generate a test case out of the PartiQL grammar file which doesn't
really validate anything about the grammar serialization other than we
get something the library would produce.