Add prototoype PEG parser using Pest.rs (#58)
This removes the previous pest scanner and replaces it with a pest
parser. Some of the rules used by the scanner conflicted with the way
rules needed to be formulated for a full parse.

The parser does _not_ parse all the way to an AST, but rather parses to
Pest's `Pairs` abstraction.