Add `AstNode`, `ToAstNode`, and refactor operators (#94)
- Adds `AstNode` with `Span`.
- Moves `partiql_parser::location` to `partiql_common::srcmap::location`.

This change adds a new structure `AstNode` structure and `ToAstNode` trait along with other related structures to enable conversion of AST structures to `AstNode` so that we can encode additional data like `Span` to an `AstNode`.

This change refactors AST operators by moving the individual structures
to Enums where applicable. It also includes the required changes to the parser.

In addition, this change moves `location` from `partiql-parser` and moves it to a package and module `partiql-common::srcmap::location`.