Add CASE expression to Parser
Adds CASE expression parsing

GitHub Issue: #119

Adds the required constructs to enable CASE EXPRESSION parsing.
The PR includes changes to partiql-ast and partiql-parser to handle
SearchedCase and SimpleSearch parsing.

See the mentioned GitHub issue for more details.

As a result of the changes in this PR, the conformance tests in the
following pass:
https://github.com/partiql/partiql-tests/blob/main/partiql-test-data/pass/parser/primitives/case.ion