Add remaining path navigation to parser
This PR includes changes for enabling Parser to parse
Path Navigations as specified in Section 4 of PartiQL
Specification. It ensures the following tests go through:
https://github.com/partiql/partiql-tests/blob/c76715ddb95ba6eec2d34ca500077b6e7eba9e35/partiql-test-data/pass/parser/primitives/path-expression.ion

In addition, the PR introduces `dbl_quoted` property to `SymbolPrimitive`
in order to cater for symbol primitives that are double quoted, e.g.
`AS "date".