Add remaining path navigation to parser
This PR includes changes for enabling Parser to parse
Path Navigations as specified in Section 4 of PartiQL
Specification. It also ensures the following tests go through:
https://github.com/partiql/partiql-tests/blob/c76715ddb95ba6eec2d34ca500077b6e7eba9e35/partiql-test-data/pass/parser/primitives/path-expression.ion

The following PR in `partiql-tests` is created for adding additional conformance
tests for path expressions:
https://github.com/partiql/partiql-tests/pull/13

In addition:
- the PR introduces `dbl_quoted` property to `SymbolPrimitive`
in order to cater for symbol primitives that are double quoted, e.g.
`AS "date".

- Splits `Identifier` to `QuotedIdentifier` and `UnquotedIdentifer`
  in order for the parser to populate the  AST with the correct
   CaseSensitivity.