[steel thread] query plan and eval using DAG model
Adds the required changes for enabling query evaluation using a DAG model.

What is our definition of DAG?
We are only interested in DAGs that can be used as execution plans, which leads to the
following definition. A DAG is a directed, cycle-free graph G = (V, E) with a denoted root node v0 ∈ V such
that all v ∈ V \{v0} are reachable from v0. Note that this is the definition of trees
without the condition |E| = |V | − 1. Hence, all trees are DAGs.
Reference: https://link.springer.com/article/10.1007/s00450-009-0061-0

In addition:
- The steel thread includes only `Scan` and `Project` operators.
- The steel thread includes logical and evaluation planning as well as interpretation (evaluation).
- The code is _not_ clean, maybe not idiomatic at some instances, and lacks documentation.
  I'll iterate through the code with a consequent PR when adding new operators.

What is the algorithm:
We're loosely using a push model with DAG as called out in the following:
https://link.springer.com/article/10.1007/s00450-009-0061-0 (Section 10.3.5. Pushing)

1. Create a DAG logical plan `LogicalPlan` with each node as a logical operator (for now manual until we integrate AST > PLAN)
2. Create an evaluation plan `EvalPlan` with the same structure as logical plan.
3. Enrich each node of `EvalPlan` with consumers of the node as `consumers`. (this could also happen at step 2).
4. Toposort the `EvalPlan` to ensure all dependencies of a node get evaluated before the node itself.
5. For each operator in `EvalPlan`, eval the node.
6. Store the result in the node as `output`.
7. Push the output to each consumer.
 For now we use a `Sink` node that is considered as the node that accumulates; we need to decide on this.

Finally, we're not performing any optimization at this stage.