Implement subquery eval using sub-plan approach
This commit includes changes to enable subquery evaluation. This is by no
means a final solution and the intention is to put a model as reference
so that we can discuss and iterate on it.

The change also includes an implementation of a LATERAL JOIN. As discussed
with Josh and Alan, putting the changes here w/ a TODO to revisit our
current JOIN implementation to cater for LATERAL functionality.

The following tests are currently passing:
- `SELECT t.a, s FROM data AS t, (SELECT v.a*2 AS u FROM t AS v) AS s;`
- `SELECT t.a, (SELECT v.a*2 AS u FROM t AS v) AS s FROM data AS t;`;