Preliminary work to create a Catalog, plumb it through lowering and evaluation with a `read_ion` extension function. (#375)
MVP for a Catalog for eventually holding Functions, Types, Schemas, etc. For now, it is able to hold a BaseTableFunction that is used to allow scanning of arbitrary data sources via an 'extension' plug-in interface.

The new partiql-extension-ion-functions crate provides an example an extension in IonExtension, which registers a ReadIonFunction.