Move `name_resolver` to a different crate
This PR is in preparation of upcomming work for adding an AST Typer.
The name resolver seems to be required by the `typer` but it being in
`partiql-logical-planner` implies its use only in planner.

In addition, this PR defines a new crate call `partiql-ast-passes`
(could pick a better name and open to ideas)  that
intends to include passes on AST transformation. Arguably `lower` in
`partiql-logical-planner` can also be part of this new crate but leaving
it out for now, as logical planning is a major transform and perhaps deserves
its own crate.