Add Catalog to NameResolver (#425)
During the work for adding a steel thread for PartiQL Typing (Recent PR #389)
and after merging #410 to `feat-type-plan-poc` it is realized that we need to
refactor the code to remove `DynamicLocalup` `VarExpr` with the assumption that
we work based off of Typing and Value Environment from the Catalog. We have a
Typing Environment in the Catalog at the moment and we are going to add the
Variable Environment as well. In preparation for such task, we need to make
the `NameResolver` Catalog aware. In that regard this commit adds the `Catalog`
to `NameResolver`

Expecting subsequent PR(s) for the name resolving using the Catalog.