Remove `NULL` and `MISSING` `partiql-types` (#463)
As we are making progress towards fleshing out partiql type semantics with more details, the current informal consensus is that NULL and MISSING can only have meaning as values or constraints (E.g., NOT NULL); considering this, this PR removes PartiQLType::NULL and PartiQLType::MISSING.

Important to note is, this PR, types literals NULL and MISSING values as Undefined (or Unknown). The rationale for making this experimental decision is that both NULL and MISSING absent values represent no types; NULL is a property of present types rather than a type itself. E.g., in SQL NULL can be assigned to any type of column. In addition, its presence in operations can lead to NULL. This follows SQL's Three-Valued Logic i.e., "If a null value affects the result of a logical expression, the result is neither true nor false but unknown.".

One could argue that the Unknown type for PartiQL absent values (NULL and MISSING) is an inhabited Bottom Types.

This PR also includes the following:

- removes unused partiql_ast_passes::partiql_typer
- fixes Clippy multiple_bound_locations errors partiql-eval